feat(backend): security, money-safety & idempotency hardening + tests/CI
Security:
- untrack committed backend/.env & frontend/.env, add .env gitignore rules
- JWT_SECRET fail-fast in production; seed passwords sourced from env
- get_current_user validates the user against the DB (is_active/soft-delete),
  switch to HTTPBearer for correct 401s + OpenAPI auth

Money integrity:
- CompanyLedger.amount typed Decimal (was float)
- race-safe atomic customer-balance updates (UPDATE ... SET balance = balance +
  delta) replacing read-modify-write, preserving every sign convention

Idempotency:
- Transaction.idempotency_key (nullable unique) + migration
- create_order/payment/discount/purchase dedupe on the client Idempotency-Key

Testing/CI:
- pytest suite (balance math, idempotency, check lifecycle, auth/RBAC) against
  a real Postgres; GitHub Actions running black/alembic/pytest + frontend build

NOTE: rotate the previously-committed JWT_SECRET/Slack webhook/DB password and
scrub git history separately (secrets remain in prior commits).

+++ backend/app/models/ledger.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 from enum import Enum as PyEnum
 
 import sqlalchemy as sa
@@ -46,7 +47,7 @@ class CompanyLedger(Base, BaseMixin):
         ),
         nullable=False,
     )
-    amount: Mapped[float] = mapped_column(sa.Numeric(12, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(sa.Numeric(12, 2), nullable=False)
     category: Mapped[str | None] = mapped_column(sa.String(100), nullable=True)
     notes: Mapped[str | None] = mapped_column(sa.Text, nullable=True)
